Some boilerplate structural placeholders.
Modes for a title screen, an "ingame" mode, and a credits scroll.  Was
a little awkward not having a credit scroll for the last GiaW because I
had no time to add it right at the end.  This time, I'm putting it in
early!

Adds some simple free fonts.

Still no real game content.